wrong handling of bit 11 (#222)
* wrong handling of bit 11

APPENDIX D - Language Encoding (EFS)
------------------------------------
...
D.2 If general purpose bit 11 is unset, the file name and comment SHOULD conform 
to the original ZIP character encoding.  If general purpose bit 11 is set, the 
filename and comment MUST support The Unicode Standard, Version 4.1.0 or 
greater using the character encoding form defined by the UTF-8 storage 
specification.  The Unicode Standard is published by the The Unicode
Consortium (www.unicode.org).  UTF-8 encoded data stored within ZIP files 
is expected to not include a byte order mark (BOM). 
...

----

For UTF-8 it is a MUST to set bit 11.
If it's not UTF-8, just don't set it and don't make it an exception

* Update ZipStreamTest.php

* cs-fix